requirements.txt
beautifulsoup4
lxml
requests
pandas
numpy
psycopg2-binary
SQLAlchemy
PyYAML
